The whole filter is the shader class now
The pass was two halves that a call site had to know how to drive: a model in dlssnr/ and a compute
shader in shaders/dlssnr/, with the orchestration living outside both. The maintainer asked for one
class dispatched like every other shader here, and he asked because he is building a reprojection
stage that presents itself as frame generation output -- somewhere with no upscaler, no NGX
parameter block, and no place to put orchestration that lives in a call site.

So it is one class. DlssNr_Dx12::Dispatch brings the model up if it is not already, builds the
feature and rebuilds it when the tuning or the resolution changes, evaluates it, and runs the three
compute passes that show it the frame and bring its answer back. The call site is one line. The
single compute dispatch it drives three times is private now, as DispatchPass.

The instance is the pass, so the caller holds it: an object cannot create itself from inside its own
method, and the call site is where the device comes from anyway.

DlssNrFeature_Dx12.cpp is gone -- its contents are the class. Its header stays as the namespace-level
API the menu and the call sites use.
